multipath: move config_reader Actor and 8to9 Models to el8toel9 repo
The config_reader actor is mostly decidated to checking for issues that
can happen during an el8 to el9 upgrade and produces
MultipathConfFacts8to9 messages. The MultipathConfig8to9 and
MultipathConfFacts8to9 models are exclusively for el8 to el9 upgrades.
All of these should live in the el8toel9 repo.

Jira: RHEL-151509
