+++ repos/system_upgrade/common/models/multipath.py
@@ -34,45 +34,3 @@ class MultipathConfigUpdatesInfo(Model):
 
     updates = fields.List(fields.Model(UpdatedMultipathConfig), default=[])
     """ Collection of multipath config updates that must be performed during the upgrade. """
-
-
-class MultipathConfig8to9(Model):
-    """
-    Model information about multipath configuration file important for the 8>9 upgrade path.
-
-    Note: This model is in the common repository due to the technical reasons
-          (reusing parser code in a single actor), and it should not be emitted on
-          non-8to9 upgrade paths. In the future, this model will likely be moved into
-          el8toel9 repository.
-    """
-    topic = SystemInfoTopic
-
-    pathname = fields.String()
-    """Config file path name"""
-
-    config_dir = fields.Nullable(fields.String())
-    """Value of config_dir in the defaults section. None if not set"""
-
-    enable_foreign_exists = fields.Boolean(default=False)
-    """True if enable_foreign is set in the defaults section"""
-
-    invalid_regexes_exist = fields.Boolean(default=False)
-    """True if any regular expressions have the value of "*" """
-
-    allow_usb_exists = fields.Boolean(default=False)
-    """True if allow_usb_devices is set in the defaults section."""
-
-
-class MultipathConfFacts8to9(Model):
-    """
-    Model representing information from multipath configuration files important for the 8>9 upgrade path.
-
-    Note: This model is in the common repository due to the technical reasons
-          (reusing parser code in a single actor), and it should not be emitted on
-          non-8to9 upgrade paths. In the future, this model will likely be moved into
-          el8toel9 repository.
-    """
-    topic = SystemInfoTopic
-
-    configs = fields.List(fields.Model(MultipathConfig8to9), default=[])
-    """List of multipath configuration files"""
