+++ repos/system_upgrade/el8toel9/actors/multipathconfread/actor.py
@@ -4,7 +4,7 @@ from leapp.models import DistributionSignedRPM, MultipathConfFacts8to9, Multipat
 from leapp.tags import FactsPhaseTag, IPUWorkflowTag
 
 
-class MultipathConfRead(Actor):
+class MultipathConfRead8to9(Actor):
     """
     Read multipath configuration files and extract the necessary information
 
@@ -13,13 +13,11 @@ class MultipathConfRead(Actor):
       - /etc/multipath/ - any files inside the directory
       - /etc/xdrdevices.conf
 
-    Two kinds of messages are generated:
-      - MultipathInfo - general information about multipath, version agnostic
-      - upgrade-path-specific messages such as MultipathConfFacts8to9 (produced only
-        when upgrading from 8 to 9)
+    Produces MultipathInfo with general information about multipath, and
+    MultipathConfFacts8to9 with details needed for the 8 to 9 upgrade.
     """
 
-    name = 'multipath_conf_read'
+    name = 'multipath_conf_read_8to9'
     consumes = (DistributionSignedRPM,)
     produces = (MultipathInfo, MultipathConfFacts8to9)
     tags = (FactsPhaseTag, IPUWorkflowTag)
